restructured portfolio folder
