+++ src/App.js
@@ -6,7 +6,7 @@ import "./App.css";
 import Layout from "./hoc/Layout/Layout";
 import SymbolSearch from "./components/SymbolSearch/SymbolSearch";
 import SymbolInfo from "./components/SymbolInfo/SymbolInfo";
-import Portfolio from "./components/Portfolio/Portfolio";
+import Portfolios from "./components/Portfolios/Portfolios";
 
 const App = (props) => {
   let routes = (
@@ -14,7 +14,7 @@ const App = (props) => {
       <Route path="/" exact render={() => <h1>Homepage Coming Soon</h1>} />
       <Route path={"/symbol"} component={SymbolInfo} />
       <Route path={"/search"} component={SymbolSearch} />
-      <Route path={"/portfolio"} component={Portfolio} />
+      <Route path={"/portfolios"} component={Portfolios} />
       <Route path={"/markets"} render={() => <h1>Markets Coming Soon</h1>} />
       <Route path={"/crypto"} render={() => <h1>Crypto Coming Soon</h1>} />
       <Route
