+++ src/components/Portfolios/PortfolioForm/PortfolioFormModal.js
@@ -1,8 +1,6 @@
 import React, { useState } from "react";
 import { useAuth0 } from "@auth0/auth0-react";
 
-import Aux from "../../../hoc/Aux/Aux";
-
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 import Modal from "react-bootstrap/Modal";
